minor refactor to TeisMapperFeatureCollection
