+++ app/src/main/java/org/dhis2/uicomponents/map/geometry/mapper/MapTeisToFeatureCollection.kt
@@ -5,6 +5,7 @@ import com.mapbox.geojson.Feature
 import com.mapbox.geojson.FeatureCollection
 import org.dhis2.uicomponents.map.geometry.bound.BoundsGeometry
 import org.dhis2.uicomponents.map.geometry.point.MapPointToFeature
+import org.dhis2.uicomponents.map.geometry.polygon.MapPolygonPointToFeature
 import org.dhis2.uicomponents.map.geometry.polygon.MapPolygonToFeature
 import org.dhis2.usescases.searchTrackEntity.adapters.SearchTeiModel
 import org.hisp.dhis.android.core.common.FeatureType
@@ -42,7 +43,6 @@ class MapTeisToFeatureCollection(private val bounds: BoundsGeometry,
             }
         }
 
-
         val featureCollectionMap = HashMap<String, FeatureCollection>()
         featureCollectionMap[TEI] =  FeatureCollection.fromFeatures(featureMap[TEI] as ArrayList)
         featureCollectionMap[ENROLLMENT] = FeatureCollection.fromFeatures(featureMap[ENROLLMENT] as ArrayList)
@@ -54,57 +54,38 @@ class MapTeisToFeatureCollection(private val bounds: BoundsGeometry,
         )
     }
 
-    private fun mapToPolygonEnrollment(geometry: Geometry, searchTeiModel: SearchTeiModel,
-                                       featureMap: HashMap<String, ArrayList<Feature>>) {
-        val polygon = mapPolygonToFeature.map(geometry, bounds)?.first
-        polygon?.addStringProperty(ENROLLMENT_UID, searchTeiModel.selectedEnrollment.uid())
-        polygon?.addStringProperty(TEI_UID, searchTeiModel.tei.uid())
-        featureMap[ENROLLMENT]!!.add(polygon!!)
-        val polygonPoint = mapPolygonPointToFeature.map(geometry)
-        polygonPoint.addStringProperty(TEI_UID, searchTeiModel.tei.uid())
-        featureMap[ENROLLMENT]!!.add(polygonPoint)
-    }
-
-    private fun mapToPointEnrollment(geometry: Geometry, searchTeiModel: SearchTeiModel,
-                                     featureMap: HashMap<String, ArrayList<Feature>>) {
-        val point = mapPointToFeature.map(geometry, bounds)?.first
-        if (point != null) {
-            point.addStringProperty(ENROLLMENT_UID, searchTeiModel.selectedEnrollment.uid())
-            point.addStringProperty(TEI_UID, searchTeiModel.tei.uid())
-            featureMap[ENROLLMENT]!!.add(point)
-        }
-    }
-
     private fun mapToPointTei(geometry: Geometry, searchTeiModel: SearchTeiModel,
                               featureMap: HashMap<String, ArrayList<Feature>>) {
-
         val point = mapPointToFeature.map(geometry, bounds)?.first
-        point?.apply {
-            addStringProperty(TEI_UID, searchTeiModel.tei.uid())
-            addStringProperty(TEI_IMAGE, searchTeiModel.profilePicturePath)
-            if (searchTeiModel.selectedEnrollment != null) {
-                addStringProperty(ENROLLMENT_UID, searchTeiModel.selectedEnrollment.uid())
-            }
-        }?.also { featureMap[TEI]!!.add(it) }
+        point?.addTeiInfo(searchTeiModel)?.also { featureMap[TEI]!!.add(it) }
     }
 
     private fun mapToPolygonTei(searchTeiModel: SearchTeiModel,
                                 featureMap: HashMap<String, ArrayList<Feature>>,
                                 geometry: Geometry) {
-
         val polygon = mapPolygonToFeature.map(geometry, bounds)?.first
-        polygon?.let {
-            it.addStringProperty(TEI_UID, searchTeiModel.tei.uid())
-            it.addStringProperty(TEI_IMAGE, searchTeiModel.profilePicturePath)
-            if (searchTeiModel.selectedEnrollment != null) {
-                it.addStringProperty(ENROLLMENT_UID, searchTeiModel.selectedEnrollment.uid())
-            }
-            featureMap[TEI]!!.add(it)
-        }
+        polygon?.addTeiInfo(searchTeiModel)?.also { featureMap[TEI]!!.add(it) }
+
         mapPolygonPointToFeature.map(geometry).apply {
             addStringProperty(TEI_UID, searchTeiModel.tei.uid())
             addStringProperty(TEI_IMAGE, searchTeiModel.profilePicturePath)
         }.also { featureMap[TEI]!!.add(it) }
+    }
+
+    private fun mapToPointEnrollment(geometry: Geometry, searchTeiModel: SearchTeiModel,
+                                     featureMap: HashMap<String, ArrayList<Feature>>) {
+        val point = mapPointToFeature.map(geometry, bounds)?.first
+        point.addTeiEnrollmentInfo(searchTeiModel)?.also { featureMap[ENROLLMENT]!!.add(it) }
+    }
+
+    private fun mapToPolygonEnrollment(geometry: Geometry, searchTeiModel: SearchTeiModel,
+                                       featureMap: HashMap<String, ArrayList<Feature>>) {
+        val polygon = mapPolygonToFeature.map(geometry, bounds)?.first
+        polygon.addTeiEnrollmentInfo(searchTeiModel)?.also { featureMap[ENROLLMENT]!!.add(polygon!!) }
+
+        val polygonPoint = mapPolygonPointToFeature.map(geometry)
+        polygonPoint.addStringProperty(TEI_UID, searchTeiModel.tei.uid())
+        featureMap[ENROLLMENT]!!.add(polygonPoint)
     }
 
     private fun teiHasCoordinates(searchTeiModel: SearchTeiModel) =
