+++ app/src/main/java/org/dhis2/usescases/searchTrackEntity/SearchTEModule.java
@@ -5,9 +5,9 @@ import androidx.annotation.NonNull;
 import org.dhis2.data.dagger.PerActivity;
 import org.dhis2.data.schedulers.SchedulerProvider;
 import org.dhis2.uicomponents.map.geometry.bound.BoundsGeometry;
-import org.dhis2.uicomponents.map.geometry.mapper.MapPolygonPointToFeature;
 import org.dhis2.uicomponents.map.geometry.mapper.MapTeisToFeatureCollection;
 import org.dhis2.uicomponents.map.geometry.point.MapPointToFeature;
+import org.dhis2.uicomponents.map.geometry.polygon.MapPolygonPointToFeature;
 import org.dhis2.uicomponents.map.geometry.polygon.MapPolygonToFeature;
 import org.dhis2.utils.analytics.AnalyticsHelper;
 import org.hisp.dhis.android.core.D2;
